Update the samples to use Host.CreateApplicationBuilder() instead of HostBuilder
