+++ samples/Contruum/Contruum.Server/Program.cs
@@ -146,10 +146,6 @@ builder.Services.AddOpenIddict()
         options.EnableAuthorizationEntryValidation();
     });
 
-// Register the worker responsible for creating and seeding the SQL database.
-// Note: in a real world application, this step should be part of a setup script.
-builder.Services.AddHostedService<Worker>();
-
 var app = builder.Build();
 
 if (builder.Environment.IsDevelopment())
@@ -167,4 +163,33 @@ app.UseAuthorization();
 
 app.MapRazorPages();
 
-app.Run();
+// Before starting the host, create the database used to store the application data.
+//
+// Note: in a real world application, this step should be part of a setup script.
+await using (var scope = app.Services.CreateAsyncScope())
+{
+    var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
+    await context.Database.EnsureCreatedAsync();
+
+    var manager = scope.ServiceProvider.GetRequiredService<IOpenIddictApplicationManager>();
+
+    // Retrieve the client definitions from the configuration
+    // and insert them in the applications table if necessary.
+    var descriptors = app.Configuration.GetSection("OpenIddict:Clients").Get<OpenIddictApplicationDescriptor[]>();
+    if (descriptors is not { Length: > 0 })
+    {
+        throw new InvalidOperationException("No client application was found in the configuration file.");
+    }
+
+    foreach (var descriptor in descriptors)
+    {
+        if (await manager.FindByClientIdAsync(descriptor.ClientId!) is not null)
+        {
+            continue;
+        }
+
+        await manager.CreateAsync(descriptor);
+    }
+}
+
+await app.RunAsync();
