+++ samples/Dantooine/Dantooine.WebAssembly.Server/Program.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+using OpenIddict.Abstractions;
 using OpenIddict.Client;
 using Quartz;
 using Yarp.ReverseProxy.Forwarder;
@@ -223,10 +224,6 @@ builder.Services.AddReverseProxy()
 // handler that will be used to attach the access tokens to HTTP requests or refresh tokens if necessary.
 builder.Services.Replace(ServiceDescriptor.Singleton<IForwarderHttpClientFactory, TokenRefreshingForwarderHttpClientFactory>());
 
-// Register the worker responsible for creating the database used to store tokens.
-// Note: in a real world application, this step should be part of a setup script.
-builder.Services.AddHostedService<Worker>();
-
 var app = builder.Build();
 
 if (builder.Environment.IsDevelopment())
@@ -259,7 +256,16 @@ app.MapReverseProxy(ConfigureProxyPipeline).DisableCookieRedirect();
 
 app.MapFallbackToPage("/_Host");
 
-app.Run();
+// Before starting the host, create the database used to store the application data.
+//
+// Note: in a real world application, this step should be part of a setup script.
+await using (var scope = app.Services.CreateAsyncScope())
+{
+    var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
+    await context.Database.EnsureCreatedAsync();
+}
+
+await app.RunAsync();
 
 static void ConfigureProxyPipeline(IReverseProxyApplicationBuilder app)
 {
