+++ samples/Hollastin/Hollastin.Server/Program.cs
@@ -80,10 +80,6 @@ builder.Services.AddOpenIddict()
         options.UseAspNetCore();
     });
 
-// Register the worker responsible for creating and seeding the SQL database.
-// Note: in a real world application, this step should be part of a setup script.
-builder.Services.AddHostedService<Worker>();
-
 var app = builder.Build();
 
 app.UseDeveloperExceptionPage();
@@ -98,4 +94,13 @@ app.MapDefaultControllerRoute();
 
 app.UseWelcomePage("/");
 
-app.Run();
+// Before starting the host, create the database used to store the application data.
+//
+// Note: in a real world application, this step should be part of a setup script.
+await using (var scope = app.Services.CreateAsyncScope())
+{
+    var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
+    await context.Database.EnsureCreatedAsync();
+}
+
+await app.RunAsync();
