--- samples/Matty/Matty.Server/Worker.cs
@@ -1,46 +1,0 @@
-using OpenIddict.Abstractions;
-using Matty.Server.Data;
-using static OpenIddict.Abstractions.OpenIddictConstants;
-
-namespace Matty.Server;
-
-public class Worker : IHostedService
-{
-    private readonly IServiceProvider _serviceProvider;
-
-    public Worker(IServiceProvider serviceProvider)
-        => _serviceProvider = serviceProvider;
-
-    public async Task StartAsync(CancellationToken cancellationToken)
-    {
-        await using var scope = _serviceProvider.CreateAsyncScope();
-
-        var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
-        await context.Database.EnsureCreatedAsync();
-
-        var manager = scope.ServiceProvider.GetRequiredService<IOpenIddictApplicationManager>();
-
-        if (await manager.FindByClientIdAsync("device") == null)
-        {
-            await manager.CreateAsync(new OpenIddictApplicationDescriptor
-            {
-                ClientId = "device",
-                ClientType = ClientTypes.Public,
-                ConsentType = ConsentTypes.Explicit,
-                DisplayName = "Device client",
-                Permissions =
-                {
-                    Permissions.GrantTypes.DeviceCode,
-                    Permissions.GrantTypes.RefreshToken,
-                    Permissions.Endpoints.DeviceAuthorization,
-                    Permissions.Endpoints.Token,
-                    Permissions.Scopes.Email,
-                    Permissions.Scopes.Profile,
-                    Permissions.Scopes.Roles,
-                }
-            });
-        }
-    }
-
-    public Task StopAsync(CancellationToken cancellationToken) => Task.CompletedTask;
-}
