+++ samples/Mimban/Mimban.Client/Program.cs
@@ -5,69 +5,72 @@ using Microsoft.Extensions.Logging;
 using Mimban.Client;
 using OpenIddict.Client;
 
-var host = new HostBuilder()
-    .ConfigureLogging(options => options.AddDebug())
-    .ConfigureServices(services =>
+var builder = Host.CreateApplicationBuilder(args);
+
+builder.Logging.ClearProviders();
+builder.Logging.AddDebug();
+
+builder.Services.AddDbContext<DbContext>(options =>
+{
+    options.UseSqlite($"Filename={Path.Combine(Path.GetTempPath(), "openiddict-mimban-client.sqlite3")}");
+    options.UseOpenIddict();
+});
+
+builder.Services.AddOpenIddict()
+
+    // Register the OpenIddict core components.
+    .AddCore(options =>
     {
-        services.AddDbContext<DbContext>(options =>
-        {
-            options.UseSqlite($"Filename={Path.Combine(Path.GetTempPath(), "openiddict-mimban-client.sqlite3")}");
-            options.UseOpenIddict();
-        });
-
-        services.AddOpenIddict()
-
-            // Register the OpenIddict core components.
-            .AddCore(options =>
-            {
-                // Configure OpenIddict to use the Entity Framework Core stores and models.
-                // Note: call ReplaceDefaultEntities() to replace the default OpenIddict entities.
-                options.UseEntityFrameworkCore()
-                       .UseDbContext<DbContext>();
-            })
-
-            // Register the OpenIddict client components.
-            .AddClient(options =>
-            {
-                // Note: this sample uses the authorization code flow,
-                // but you can enable the other flows if necessary.
-                options.AllowAuthorizationCodeFlow()
-                       .AllowRefreshTokenFlow();
-
-                // Register the signing and encryption credentials used to protect
-                // sensitive data like the state tokens produced by OpenIddict.
-                options.AddDevelopmentEncryptionCertificate()
-                       .AddDevelopmentSigningCertificate();
-
-                // Add the operating system integration.
-                options.UseSystemIntegration();
-
-                // Register the System.Net.Http integration and use the identity of the current
-                // assembly as a more specific user agent, which can be useful when dealing with
-                // providers that use the user agent as a way to throttle requests (e.g Reddit).
-                options.UseSystemNetHttp()
-                       .SetProductInformation(typeof(Program).Assembly);
-
-                // Add a client registration matching the client application definition in the server project.
-                options.AddRegistration(new OpenIddictClientRegistration
-                {
-                    Issuer = new Uri("https://localhost:44383/", UriKind.Absolute),
-
-                    ClientId = "console_app",
-                    RedirectUri = new Uri("/", UriKind.Relative)
-                });
-            });
-
-        // Register the worker responsible for creating the database used to store tokens
-        // and adding the registry entries required to register the custom URI scheme.
-        //
-        // Note: in a real world application, this step should be part of a setup script.
-        services.AddHostedService<Worker>();
-
-        // Register the background service responsible for handling the console interactions.
-        services.AddHostedService<InteractiveService>();
+        // Configure OpenIddict to use the Entity Framework Core stores and models.
+        // Note: call ReplaceDefaultEntities() to replace the default OpenIddict entities.
+        options.UseEntityFrameworkCore()
+               .UseDbContext<DbContext>();
     })
-    .UseConsoleLifetime()
-    .Build();
 
-await host.RunAsync();
+    // Register the OpenIddict client components.
+    .AddClient(options =>
+    {
+        // Note: this sample uses the authorization code flow,
+        // but you can enable the other flows if necessary.
+        options.AllowAuthorizationCodeFlow()
+               .AllowRefreshTokenFlow();
+
+        // Register the signing and encryption credentials used to protect
+        // sensitive data like the state tokens produced by OpenIddict.
+        options.AddDevelopmentEncryptionCertificate()
+               .AddDevelopmentSigningCertificate();
+
+        // Add the operating system integration.
+        options.UseSystemIntegration();
+
+        // Register the System.Net.Http integration and use the identity of the current
+        // assembly as a more specific user agent, which can be useful when dealing with
+        // providers that use the user agent as a way to throttle requests (e.g Reddit).
+        options.UseSystemNetHttp()
+               .SetProductInformation(typeof(Program).Assembly);
+
+        // Add a client registration matching the client application definition in the server project.
+        options.AddRegistration(new OpenIddictClientRegistration
+        {
+            Issuer = new Uri("https://localhost:44383/", UriKind.Absolute),
+
+            ClientId = "console_app",
+            RedirectUri = new Uri("/", UriKind.Relative)
+        });
+    });
+
+// Register the background service responsible for handling the console interactions.
+builder.Services.AddHostedService<InteractiveService>();
+
+var app = builder.Build();
+
+// Before starting the host, create the database used to store the application data.
+//
+// Note: in a real world application, this step should be part of a setup script.
+await using (var scope = app.Services.CreateAsyncScope())
+{
+    var context = scope.ServiceProvider.GetRequiredService<DbContext>();
+    await context.Database.EnsureCreatedAsync();
+}
+
+await app.RunAsync();
