+++ samples/Sorgan/Sorgan.Console.Client/Program.cs
@@ -4,81 +4,78 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Sorgan.Console.Client;
 
-var host = new HostBuilder()
-    // Note: applications for which a single instance is preferred can reference
-    // the Dapplo.Microsoft.Extensions.Hosting.AppServices package and call this
-    // method to automatically close extra instances based on the specified identifier:
-    //
-    // .ConfigureSingleInstance(options => options.MutexId = "{5519A32F-5B86-4CBB-A601-0CC7872A126A}")
-    //
-    .ConfigureLogging(options => options.AddDebug())
-    .ConfigureServices(services =>
+var builder = Host.CreateApplicationBuilder(args);
+
+builder.Logging.ClearProviders();
+builder.Logging.AddDebug();
+
+builder.Services.AddDbContext<DbContext>(options =>
+{
+    options.UseSqlite($"Filename={Path.Combine(Path.GetTempPath(), "openiddict-sorgan-console-client.sqlite3")}");
+    options.UseOpenIddict();
+});
+
+builder.Services.AddOpenIddict()
+
+    // Register the OpenIddict core components.
+    .AddCore(options =>
     {
-        services.AddDbContext<DbContext>(options =>
-        {
-            options.UseSqlite($"Filename={Path.Combine(Path.GetTempPath(), "openiddict-sorgan-console-client.sqlite3")}");
-            options.UseOpenIddict();
-        });
-
-        services.AddOpenIddict()
-
-            // Register the OpenIddict core components.
-            .AddCore(options =>
-            {
-                // Configure OpenIddict to use the Entity Framework Core stores and models.
-                // Note: call ReplaceDefaultEntities() to replace the default OpenIddict entities.
-                options.UseEntityFrameworkCore()
-                       .UseDbContext<DbContext>();
-            })
-
-            // Register the OpenIddict client components.
-            .AddClient(options =>
-            {
-                // Note: this sample uses the authorization code, device authorization code
-                // and refresh token flows, but you can enable the other flows if necessary.
-                options.AllowAuthorizationCodeFlow()
-                       .AllowDeviceAuthorizationFlow()
-                       .AllowRefreshTokenFlow();
-
-                // Register the signing and encryption credentials used to protect
-                // sensitive data like the state tokens produced by OpenIddict.
-                options.AddDevelopmentEncryptionCertificate()
-                       .AddDevelopmentSigningCertificate();
-
-                // Add the operating system integration.
-                options.UseSystemIntegration();
-
-                // Register the System.Net.Http integration and use the identity of the current
-                // assembly as a more specific user agent, which can be useful when dealing with
-                // providers that use the user agent as a way to throttle requests (e.g Reddit).
-                options.UseSystemNetHttp()
-                       .SetProductInformation(typeof(Program).Assembly);
-
-                // Register the Web providers integrations.
-                //
-                // Note: to mitigate mix-up attacks, it's recommended to use a unique redirection endpoint
-                // address per provider, unless all the registered providers support returning an "iss"
-                // parameter containing their URL as part of authorization responses. For more information,
-                // see https://datatracker.ietf.org/doc/html/draft-ietf-oauth-security-topics#section-4.4.
-                options.UseWebProviders()
-                       .AddGitHub(options =>
-                       {
-                           options.SetClientId("5c11c030ca570e8c5a16")
-                                  .SetClientSecret("a5d36464b2ac2fe3e87fbfb95f0ebcf06c5992c1")
-                                  .SetRedirectUri("callback/login/github");
-                       });
-            });
-
-        // Register the worker responsible for creating the database used to store tokens
-        // and adding the registry entries required to register the custom URI scheme.
-        //
-        // Note: in a real world application, this step should be part of a setup script.
-        services.AddHostedService<Worker>();
-
-        // Register the background service responsible for handling the console interactions.
-        services.AddHostedService<InteractiveService>();
+        // Configure OpenIddict to use the Entity Framework Core stores and models.
+        // Note: call ReplaceDefaultEntities() to replace the default OpenIddict entities.
+        options.UseEntityFrameworkCore()
+               .UseDbContext<DbContext>();
     })
-    .UseConsoleLifetime()
-    .Build();
 
-await host.RunAsync();
+    // Register the OpenIddict client components.
+    .AddClient(options =>
+    {
+        // Note: this sample uses the authorization code, device authorization code
+        // and refresh token flows, but you can enable the other flows if necessary.
+        options.AllowAuthorizationCodeFlow()
+               .AllowDeviceAuthorizationFlow()
+               .AllowRefreshTokenFlow();
+
+        // Register the signing and encryption credentials used to protect
+        // sensitive data like the state tokens produced by OpenIddict.
+        options.AddDevelopmentEncryptionCertificate()
+               .AddDevelopmentSigningCertificate();
+
+        // Add the operating system integration.
+        options.UseSystemIntegration();
+
+        // Register the System.Net.Http integration and use the identity of the current
+        // assembly as a more specific user agent, which can be useful when dealing with
+        // providers that use the user agent as a way to throttle requests (e.g Reddit).
+        options.UseSystemNetHttp()
+               .SetProductInformation(typeof(Program).Assembly);
+
+        // Register the Web providers integrations.
+        //
+        // Note: to mitigate mix-up attacks, it's recommended to use a unique redirection endpoint
+        // address per provider, unless all the registered providers support returning an "iss"
+        // parameter containing their URL as part of authorization responses. For more information,
+        // see https://datatracker.ietf.org/doc/html/draft-ietf-oauth-security-topics#section-4.4.
+        options.UseWebProviders()
+                .AddGitHub(options =>
+                {
+                    options.SetClientId("5c11c030ca570e8c5a16")
+                           .SetClientSecret("a5d36464b2ac2fe3e87fbfb95f0ebcf06c5992c1")
+                           .SetRedirectUri("callback/login/github");
+                });
+    });
+
+// Register the background service responsible for handling the console interactions.
+builder.Services.AddHostedService<InteractiveService>();
+
+var app = builder.Build();
+
+// Before starting the host, create the database used to store the application data.
+//
+// Note: in a real world application, this step should be part of a setup script.
+await using (var scope = app.Services.CreateAsyncScope())
+{
+    var context = scope.ServiceProvider.GetRequiredService<DbContext>();
+    await context.Database.EnsureCreatedAsync();
+}
+
+await app.RunAsync();
