--- samples/Velusia/Velusia.Client/Worker.cs
@@ -1,21 +1,0 @@
-﻿using Velusia.Client.Models;
-
-namespace Velusia.Client;
-
-public class Worker : IHostedService
-{
-    private readonly IServiceProvider _serviceProvider;
-
-    public Worker(IServiceProvider serviceProvider)
-        => _serviceProvider = serviceProvider;
-
-    public async Task StartAsync(CancellationToken cancellationToken)
-    {
-        await using var scope = _serviceProvider.CreateAsyncScope();
-
-        var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
-        await context.Database.EnsureCreatedAsync(cancellationToken);
-    }
-
-    public Task StopAsync(CancellationToken cancellationToken) => Task.CompletedTask;
-}
