--- samples/Velusia/Velusia.Server/Worker.cs
@@ -1,59 +1,0 @@
-using OpenIddict.Abstractions;
-using Velusia.Server.Data;
-using static OpenIddict.Abstractions.OpenIddictConstants;
-
-namespace Velusia.Server;
-
-public class Worker : IHostedService
-{
-    private readonly IServiceProvider _serviceProvider;
-
-    public Worker(IServiceProvider serviceProvider)
-        => _serviceProvider = serviceProvider;
-
-    public async Task StartAsync(CancellationToken cancellationToken)
-    {
-        await using var scope = _serviceProvider.CreateAsyncScope();
-
-        var context = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
-        await context.Database.EnsureCreatedAsync();
-
-        var manager = scope.ServiceProvider.GetRequiredService<IOpenIddictApplicationManager>();
-
-        if (await manager.FindByClientIdAsync("mvc") == null)
-        {
-            await manager.CreateAsync(new OpenIddictApplicationDescriptor
-            {
-                ClientId = "mvc",
-                ClientSecret = "901564A5-E7FE-42CB-B10D-61EF6A8F3654",
-                ConsentType = ConsentTypes.Explicit,
-                DisplayName = "MVC client application",
-                RedirectUris =
-                {
-                    new Uri("https://localhost:44338/callback/login/local")
-                },
-                PostLogoutRedirectUris =
-                {
-                    new Uri("https://localhost:44338/callback/logout/local")
-                },
-                Permissions =
-                {
-                    Permissions.Endpoints.Authorization,
-                    Permissions.Endpoints.EndSession,
-                    Permissions.Endpoints.Token,
-                    Permissions.GrantTypes.AuthorizationCode,
-                    Permissions.ResponseTypes.Code,
-                    Permissions.Scopes.Email,
-                    Permissions.Scopes.Profile,
-                    Permissions.Scopes.Roles
-                },
-                Requirements =
-                {
-                    Requirements.Features.ProofKeyForCodeExchange
-                }
-            });
-        }
-    }
-
-    public Task StopAsync(CancellationToken cancellationToken) => Task.CompletedTask;
-}
