--- samples/Zirku/Zirku.Client1/Worker.cs
@@ -1,23 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-
-namespace Zirku.Client1;
-
-public class Worker : IHostedService
-{
-    private readonly IServiceProvider _provider;
-
-    public Worker(IServiceProvider provider)
-        => _provider = provider;
-
-    public async Task StartAsync(CancellationToken cancellationToken)
-    {
-        using var scope = _provider.CreateScope();
-
-        var context = scope.ServiceProvider.GetRequiredService<DbContext>();
-        await context.Database.EnsureCreatedAsync();
-    }
-
-    public Task StopAsync(CancellationToken cancellationToken) => Task.CompletedTask;
-}
